[ion] Sharing millis & micros for blackbox and simulator
