+++ ion/include/ion/timing.h
@@ -11,6 +11,9 @@ namespace Ion {
 EXTERNC uint32_t millis();
 EXTERNC uint32_t micros();
 
+EXTERNC void msleep(long ms);
+EXTERNC void usleep(long us);
+
 #ifdef __cplusplus
 }
 #endif
