+++ ion/src/emscripten/Makefile
@@ -10,6 +10,7 @@ objs += $(addprefix ion/src/shared/, \
   events_modifier.o \
   power.o \
   random.o \
+  timing.o \
   dummy/backlight.o \
   dummy/battery.o \
   dummy/fcc_id.o \
